core: AddDiskCommand - add option to leave disk as illegal on failure
When adding a disk in case of failure the added disk is currently being
removed automatically by the engine which is wrong in atleast some of
the cases as we might have garbage on the storage.
As first step, this change allows to restrict the command to leave the
disk in illegal status in case of failure and not delete the disk.

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/AddImageFromScratchCommand.java
@@ -23,6 +23,7 @@ public class AddImageFromScratchCommand<T extends AddImageFromScratchParameters>
     public AddImageFromScratchCommand(T parameters) {
         super(parameters);
         setVmId(getParameters().getMasterVmId());
+        getParameters().setCommandType(getActionType());
     }
 
     protected AddImageFromScratchCommand(Guid commandId) {
@@ -65,11 +66,19 @@ public class AddImageFromScratchCommand<T extends AddImageFromScratchParameters>
         TransactionSupport.executeInNewTransaction(new TransactionMethod<Void>() {
             @Override
             public Void runInTransaction() {
-                addDiskImageToDb(mNewCreatedDiskImage, getCompensationContext());
+                if (!getParameters().isShouldRemainIllegalOnFailedExecution()) {
+                    addDiskImageToDb(mNewCreatedDiskImage, getCompensationContext());
+                } else {
+                    addDiskImageToDb(mNewCreatedDiskImage, null);
+                    getCompensationContext().snapshotEntityStatus(mNewCreatedDiskImage.getImage(), ImageStatus.ILLEGAL);
+                }
                 return null;
             }
         });
         freeLock();
+        if (getParameters().isShouldRemainIllegalOnFailedExecution()) {
+            getReturnValue().setActionReturnValue(mNewCreatedDiskImage);
+        }
         processImageInIrs();
         getReturnValue().setActionReturnValue(mNewCreatedDiskImage);
         setSucceeded(true);
@@ -108,8 +117,13 @@ public class AddImageFromScratchCommand<T extends AddImageFromScratchParameters>
     @Override
     protected void endWithFailure() {
         if (getDestinationDiskImage() != null) {
-            DbFacade.getInstance().getDiskImageDynamicDao().remove(getDestinationDiskImage().getImageId());
+            if (getParameters().isShouldRemainIllegalOnFailedExecution()) {
+                setImageStatus(ImageStatus.ILLEGAL);
+            } else {
+                DbFacade.getInstance().getDiskImageDynamicDao().remove(getDestinationDiskImage().getImageId());
+                super.endWithFailure();
+            }
         }
-        super.endWithFailure();
+
     }
 }
